requirements.txt
Babel==2.3.4
Django==1.9.12
Faker==0.7.7
Pillow==4.0.0
Unidecode==0.04.20
django-appconf==1.0.1
django-braces==1.8.1
django-compressor==1.5
django-extra-views==0.6.4
django-haystack==2.5.1
django-oscar==1.3
django-tables2==1.0.7
django-treebeard==4.1.0
django-widget-tweaks==1.4.1
factory-boy==2.8.1
fake-factory==0.5.10
funcsigs==1.0.2
ipaddress==1.0.18
mock==2.0.0
olefile==0.44
pbr==1.10.0
phonenumbers==7.7.5
purl==1.3
pycountry==1.20
pycrypto==2.6.1
python-dateutil==2.6.0
pytz==2016.6.1
requests==2.9.1
six==1.10.0
sorl-thumbnail==12.4a1
wsgiref==0.1.2
